backend/requirements.txt
asyncpg
fastapi
uvicorn
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
numpy
opencv-python
cvzone
